+++ static/css/indexStyle.css
@@ -1,214 +1,41 @@
-:root {
-    /* html에서 :root는 html 선택자랑 같다 */
-    /* 변수처럼 사용 가능. 예시) color: var(--surface-color) */
-    --surface-color: #fff;
-    --curve: 40;
+body {
+    background-color: RGBA(232, 52, 78, 0.2);
+    min-height: 100vh;
 }
 
-.near-cards {
-    position: relative;
-    margin-top: 80px;
-}
-
-.trip-cards {
-    position: relative;
-    margin-top: 80px;
-    margin-bottom: 70px;
-}
-
-.wrap {
-    width: 80%;
+.section {
+    padding: 1rem 1.5rem;
+    max-width: 750px;
     margin: auto;
 }
 
-.slide-title {
-    font-family: 'yg-jalnan', verdana, tahoma;
-    margin-bottom: 20px;
+.title {
+    font-family: 'Jua', sans-serif;
+    font-weight: 800;
+    font-size: 5rem;
 }
 
-.slide-title a {
-    font-family: 'Noto Sans JP', sans-serif;
-    font-size: 13px;
-    color: #073A96;
-    text-decoration: none;
+.subtitle {
+    font-family: 'Nanum Gothic', sans-serif;
+    font-size: 2rem;
 }
 
-.slide-title a:hover {
-    color: #6E85B2;
+.is-sparta {
+    color: #e8344e !important;
 }
 
-.before-render {
-    position: relative;
-    width: 95%;
-    height: 300px;
-    margin: auto;
-    background-color: #EEEEEE;
-    border-radius: calc(var(--curve) * 1px);
+.button.is-sparta {
+    background-color: #e8344e;
+    border-color: transparent;
+    color: #fff !important;
 }
 
-#btn_prev {
-    position: absolute;
-    top: 50%;
-    left: 4%;
+.button.is-sparta.is-outlined {
     background-color: transparent;
-    color: #A69F99;
-    font-size: 30px;
-    border: none;
+    border-color: #e8344e;
+    color: #e8344e !important;
 }
 
-#btn_next {
-    position: absolute;
-    top: 50%;
-    right: 4%;
-    background-color: transparent;
-    color: #A69F99;
-    font-size: 30px;
-    border: none;
-}
-
-#btn_prev:hover {
-    color: #073A96;
-}
-
-#btn_next:hover {
-    color: #073A96;
-}
-
-#btn_prev2 {
-    position: absolute;
-    top: 50%;
-    left: 4%;
-    background-color: transparent;
-    color: #A69F99;
-    font-size: 30px;
-    border: none;
-}
-
-#btn_next2 {
-    position: absolute;
-    top: 50%;
-    right: 4%;
-    background-color: transparent;
-    color: #A69F99;
-    font-size: 30px;
-    border: none;
-}
-
-#btn_prev2:hover {
-    color: #073A96;
-}
-
-#btn_next2:hover {
-    color: #073A96;
-}
-
-.cards {
-    width: 100%;
-    height: 300px;
-    margin: auto;
-    padding: 0;
-    list-style-type: none;
-}
-
-.card {
-    position: relative;
-    display: block;
-    height: 100%;
-    border-radius: calc(var(--curve) * 1px);
-    overflow: hidden;
-    text-decoration: none;
-
-    box-shadow: 1px 0 3px #bbb;
-}
-
-.card__image {
-    width: 100%;
-    height: 300px;
-
-    border-radius: calc(var(--curve) * 1px);
-    /*대체 콘텐츠의 가로세로비를 유지하면서, 요소 콘텐츠 박스를 가득 채운다. 서로의 가로세로비가 일치하지 않으면 객체 일부가 잘려나간다.*/
-    object-fit: cover;
-    border: solid 2px #fef8f8;
-}
-
-.card__overlay {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    z-index: 1;
-    border-radius: calc(var(--curve) * 1px);
-    background-color: var(--surface-color);
-    transform: translateY(100%);
-    transition: .2s ease-in-out;
-}
-
-.card:hover .card__overlay {
-    transform: translateY(0);
-}
-
-.card__header {
-    position: relative;
-    display: flex;
-    align-items: center;
-    gap: 2em;
-    padding: 1em;
-    border-radius: calc(var(--curve) * 1px) 0 0 0;
-    background-color: var(--surface-color);
-    transform: translateY(-100%);
-    transition: .2s ease-in-out;
-}
-
-.card__arc {
-    width: 80px;
-    height: 80px;
-    position: absolute;
-    bottom: 100%;
-    right: 0;
-    z-index: 1;
-}
-
-.card__arc path {
-    fill: var(--surface-color);
-    d: path("M 40 80 c 22 0 40 -22 40 -40 v 40 Z");
-}
-
-.card:hover .card__header {
-    transform: translateY(0);
-}
-
-.card__thumb {
-    flex-shrink: 0;
-    width: 50px;
-    height: 50px;
-    border-radius: 50%;
-}
-
-.card__title {
-    font-size: 1em;
-    margin: 0 0 .3em;
-    color: #1B3A6A;
-}
-
-.card__status {
-    font-size: .8em;
-    color: #528BC6;
-}
-
-.card__description {
-    padding: 0 2em 2em;
-    margin: 0;
-    color: #1B3A6A;
-    font-family: "MockFlowFont";
-    font-size: 0.9em;
-    display: -webkit-box;
-    -webkit-box-orient: vertical;
-    -webkit-line-clamp: 3;
-    overflow: hidden;
-}
-
-.footer {
-    width: 100%;
-    height: 200px;
-    background-color: #252226;
+.help {
+    color: gray;
 }
